index: adjust code in preparation for large indexes
The old code was written with care to avoid problems
with >2GB indexes on 32-bit machines, using uint32
offsets throughout. But then it mmap'ed the index file,
which wouldn't work with a >2GB index anyway, so none
of that code was really useful in the end.

Move the code to use int for offsets and values throughout.
This will still not work well with >2GB indexes on 32-bit
machines, but again the code already doesn't work in that case.
Now it may at least diagnose the problem a bit better.

The old uint32 code did work well for up to 4GB indexes
on 64-bit machines. This is in preparation for >4GB indexes.

+++ index/merge_test.go
@@ -72,7 +72,7 @@ func TestMerge(t *testing.T) {
 
 	checkFiles := func(ix *Index, l ...string) {
 		for i, s := range l {
-			if n := ix.Name(uint32(i)); n != s {
+			if n := ix.Name(i); n != s {
 				t.Errorf("%s: Name(%d) = %s, want %s", nameof(ix), i, n, s)
 			}
 		}
@@ -82,7 +82,7 @@ func TestMerge(t *testing.T) {
 	checkFiles(ix2, "/b/www", "/b/xx", "/b/yy", "/cc")
 	checkFiles(ix3, "/a/x", "/a/y", "/b/www", "/b/xx", "/b/yy", "/c/ab", "/c/de", "/cc")
 
-	check := func(ix *Index, trig string, l ...uint32) {
+	check := func(ix *Index, trig string, l ...int) {
 		l1 := ix.PostingList(tri(trig[0], trig[1], trig[2]))
 		if !equalList(l1, l) {
 			t.Errorf("PostingList(%s, %s) = %v, want %v", nameof(ix), trig, l1, l)
